feat : fixed the UI of leaderboard and statistics (#125) - formatTime function created in utils and used in sidebar and leaderboard - rounded number - win rate in percent - populate stats removed
