+++ views/StatisticsView.vue
@@ -34,9 +34,6 @@ const props = defineProps({
   },
 })
 
-// Emits
-const emit = defineEmits(['populate-stats'])
-
 // Refs
 const isStatOpen = ref(false)
 
@@ -60,11 +57,34 @@ const user = useCurrentUser()
         </div>
       </template>
         <div style="max-height: 80vh; overflow-y: auto">
-          <div class="flex flex-col items-center mb-10">
-            <UButton @click="emit('populate-stats')" class="text-center">Populate stats</UButton>
-          </div>
           <div class="flex flex-col md:flex-row justify-around mt-3">
             <div class="flex justify-around w-full md:w-1/2 mb-5 md:mb-0">
+              <div class="flex flex-col items-center">
+                <div class="flex justify-center items-center h-24 w-24 rounded-full border-primary border-8 text-2xl font-extrabold">
+                  {{ averageGuesses }}
+                </div>
+                <p class="mt-3 text-sm text-gray-500">Avg. Guesses</p>
+              </div>
+              <div class="flex flex-col items-center">
+                <div class="flex justify-center items-center h-24 w-24 rounded-full border-primary border-8 text-2xl font-extrabold">
+                  {{ averageRank }}
+                </div>
+                <p class="mt-3 text-sm text-gray-500">Avg. Rank</p>
+              </div>
+              <div class="flex flex-col items-center">
+                <div class="flex justify-center items-center h-24 w-24 rounded-full border-primary border-8 text-2xl font-extrabold">
+                  {{ winRate  }}
+                </div>
+                <p class="mt-3 text-sm text-gray-500">Win Rate</p>
+              </div>
+            </div>
+            <div class="flex justify-around w-full md:w-1/2">
+              <div class="flex flex-col items-center">
+                <div class="flex justify-center items-center h-24 w-24 rounded-full border-primary border-8 text-2xl font-extrabold">
+                  {{ gamesPlayed}}
+                </div>
+                <p class="mt-3 text-sm text-gray-500">Games Played</p>
+              </div>
               <div class="flex flex-col items-center">
                 <div class="flex justify-center items-center h-24 w-24 rounded-full border-primary border-8 text-2xl font-extrabold">
                   {{ currentStreak }}
@@ -76,32 +96,6 @@ const user = useCurrentUser()
                   {{ maxStreak }}
                 </div>
                 <p class="mt-3 text-sm text-gray-500">Max. Streak</p>
-              </div>
-              <div class="flex flex-col items-center">
-                <div class="flex justify-center items-center h-24 w-24 rounded-full border-primary border-8 text-2xl font-extrabold">
-                  {{ averageRank }}
-                </div>
-                <p class="mt-3 text-sm text-gray-500">Avg. Rank</p>
-              </div>
-            </div>
-            <div class="flex justify-around w-full md:w-1/2">
-              <div class="flex flex-col items-center">
-                <div class="flex justify-center items-center h-24 w-24 rounded-full border-primary border-8 text-2xl font-extrabold">
-                  {{ averageGuesses }}
-                </div>
-                <p class="mt-3 text-sm text-gray-500">Avg. Guesses</p>
-              </div>
-              <div class="flex flex-col items-center">
-                <div class="flex justify-center items-center h-24 w-24 rounded-full border-primary border-8 text-2xl font-extrabold">
-                  {{ winRate }}
-                </div>
-                <p class="mt-3 text-sm text-gray-500">Win Rate</p>
-              </div>
-              <div class="flex flex-col items-center">
-                <div class="flex justify-center items-center h-24 w-24 rounded-full border-primary border-8 text-2xl font-extrabold">
-                  {{ gamesPlayed }}
-                </div>
-                <p class="mt-3 text-sm text-gray-500">Games Played</p>
               </div>
             </div>
           </div>
